# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Welcome the App Functions SDK for EdgeX. This sdk is meant to provide all the plumbing necessary for developers to get started\nin processing/transforming/exporting data out of EdgeX. For full documentation on this SDK please visit https://github.com/tuanldchainos/app-functions-sdk-go/blob/master/README.md\n\nThis Swagger Document focuses on the few restful APIs provided with the SDK, however the complete documentation for all available features can be found\nhere https://github.com/tuanldchainos/app-functions-sdk-go/blob/master/README.md .",
    "title": "App Functions SDK API.",
    "contact": {
      "url": "https://edgex.discourse.group/c/applications"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://github.com/tuanldchainos/app-functions-sdk-go/blob/master/LICENSE"
    },
    "version": "1.0.0"
  },
  "host": "localhost",
  "basePath": "/api/v1/",
  "paths": {
...
    "/trigger": {
      "post": {
        "description": "Available when HTTPTrigger is specified as the binding in configuration. This API\nprovides a way to initiate and start processing the defined pipeline using the data submitted.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Trigger"
        ],
        "summary": "Trigger",
        "operationId": "Trigger",
        "parameters": [
          {
            "description": "This is the data that will processed the configured pipeline. Typically this is an EdgeX event as described below, however, it can\ningest other forms of data if a custom Target Type (https://github.com/tuanldchainos/app-functions-sdk-go/blob/master/README.md#target-type) is being used.\n",
            "name": "Data Event",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get current version",
            "schema": {
              "$ref": "#/definitions/Version"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Event": {
      "description": "Event represents a single measurable event read from a device",
      "type": "object",
      "properties": {
        "created": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Created"
        },
        "device": {
          "type": "string",
          "x-go-name": "Device"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "modified": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Modified"
        },
        "origin": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Origin"
        },
        "pushed": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Pushed"
        },
        "readings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Reading"
          },
          "x-go-name": "Readings"
        }
      },
      "x-go-package": "github.com/edgexfoundry/go-mod-core-contracts/models"
    },
...
    "Reading": {
      "type": "object",
      "properties": {
        "binaryValue": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "uint8"
          },
          "x-go-name": "BinaryValue"
        },
        "created": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Created"
        },
        "device": {
          "type": "string",
          "x-go-name": "Device"
        },
        "id": {
          "type": "string",
          "x-go-name": "Id"
        },
        "modified": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Modified"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "origin": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Origin"
        },
        "pushed": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Pushed"
        },
        "value": {
          "type": "string",
          "x-go-name": "Value"
        }
      },
      "x-go-package": "github.com/edgexfoundry/go-mod-core-contracts/models"
    },
...
    "Version": {
      "type": "object",
      "properties": {
        "sdk_version": {
          "type": "string",
          "x-go-name": "SDKVersion"
        },
        "version": {
          "type": "string",
          "x-go-name": "Version"
        }
      },
      "x-go-package": "github.com/tuanldchainos/app-functions-sdk-go/internal/webserver"
    },
...
  }
}